public-osint-and-intel-tools-main/requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
python-whois==0.8.0
dnspython==2.4.2
shodan==1.31.0
tweepy==4.14.0
aiohttp==3.9.1
validators==0.22.0
colorama==0.4.6
tabulate==0.9.0
python-dotenv==1.0.0
Pillow==10.1.0
